+++ tests/testthat/test-utils-internal.R
@@ -43,10 +43,3 @@ test_that("Column output as expected", {
   expect_equal(is_separate("3,5 = age,edu", sep = "=", keep = 2), "age,edu")
 })
 
-test_that("Verbose message", {
-  op <- options()
-  options(orgdata.verbose = TRUE)
-  expect_message(is_verbose("C:/File/Name", "File:"), "File: C:/File/Name")
-  expect_message(is_verbose("C:/File/Name"), " C:/File/Name")
-  options(op)
-})